Refactor: create ccache_compression lib
