+++ src/ccache.cpp
@@ -22,15 +22,10 @@
 #include "Args.hpp"
 #include "ArgsInfo.hpp"
 #include "Checksum.hpp"
-#include "Compression.hpp"
 #include "Context.hpp"
 #include "Depfile.hpp"
-#include "core/Fd.hpp"
-#include "core/File.hpp"
 #include "Finalizer.hpp"
 #include "Hash.hpp"
-#include "core/Lockfile.hpp"
-#include "core/Logging.hpp"
 #include "Manifest.hpp"
 #include "MiniTrace.hpp"
 #include "ProgressBar.hpp"
@@ -39,18 +34,23 @@
 #include "ResultExtractor.hpp"
 #include "ResultRetriever.hpp"
 #include "SignalHandler.hpp"
-#include "core/TemporaryFile.hpp"
 #include "UmaskScope.hpp"
 #include "Util.hpp"
 #include "argprocessing.hpp"
 #include "cleanup.hpp"
 #include "compopt.hpp"
 #include "compress.hpp"
+#include "compression/Compression.hpp"
+#include "core/Fd.hpp"
+#include "core/File.hpp"
 #include "core/FormatNonstdStringView.hpp"
+#include "core/Lockfile.hpp"
+#include "core/Logging.hpp"
 #include "core/Sloppy.hpp"
 #include "core/StdMakeUnique.hpp"
-#include "core/fmtmacros.hpp"
+#include "core/TemporaryFile.hpp"
 #include "core/exceptions.hpp"
+#include "core/fmtmacros.hpp"
 #include "execute.hpp"
 #include "hashutil.hpp"
 #include "language.hpp"
