+++ src/compress.cpp
@@ -18,17 +18,17 @@
 
 #include "compress.hpp"
 
-#include "core/AtomicFile.hpp"
 #include "CacheEntryReader.hpp"
 #include "CacheEntryWriter.hpp"
 #include "Context.hpp"
-#include "core/File.hpp"
-#include "core/Logging.hpp"
 #include "Manifest.hpp"
 #include "Result.hpp"
-#include "core/Statistics.hpp"
 #include "ThreadPool.hpp"
-#include "ZstdCompressor.hpp"
+#include "compression/ZstdCompressor.hpp"
+#include "core/AtomicFile.hpp"
+#include "core/File.hpp"
+#include "core/Logging.hpp"
+#include "core/Statistics.hpp"
 #include "core/StdMakeUnique.hpp"
 #include "core/assertions.hpp"
 #include "core/fmtmacros.hpp"
