+++ src/compress.hpp
@@ -20,7 +20,7 @@
 
 #include "core/system.hpp"
 
-#include "Util.hpp"
+#include "core/Util.hpp"
 
 #include "third_party/nonstd/optional.hpp"
