+++ src/compression/src/ZstdCompressor.cpp
@@ -16,7 +16,7 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "ZstdCompressor.hpp"
+#include "compression/ZstdCompressor.hpp"
 
 #include "core/Logging.hpp"
 #include "core/assertions.hpp"
