+++ src/compression/src/ZstdDecompressor.cpp
@@ -16,7 +16,7 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "ZstdDecompressor.hpp"
+#include "compression/ZstdDecompressor.hpp"
 
 #include "core/assertions.hpp"
 #include "core/exceptions.hpp"
