+++ unittest/test_Compression.cpp
@@ -16,7 +16,7 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "../src/Compression.hpp"
+#include "compression/Compression.hpp"
 #include "core/Config.hpp"
 
 #include "third_party/doctest.h"
